feat: added common CopyValue component across Sistent color, typography and spacing pages
